Add `SpiDisplay` base class
